Move puppeteer inject helper out of createTest
