+++ test/e2e/lib/framework/index.ts
@@ -1,7 +1,8 @@
-export { createTest, injectRumWithPuppeteer } from './createTest'
+export { createTest } from './createTest'
 export { bundleSetup, html } from './pageSetups'
 export { IntakeRegistry } from './intakeRegistry'
 export { getTestServers, waitForServersIdle } from './httpServers'
 export { flushEvents } from './flushEvents'
 export { waitForRequests } from './waitForRequests'
 export { LARGE_RESPONSE_MIN_BYTE_SIZE } from './serverApps/mock'
+export { RUM_BUNDLE } from './sdkBuilds'
