+++ test/e2e/scenario/rum/s8sInject.scenario.ts
@@ -1,4 +1,7 @@
-import { injectRumWithPuppeteer } from '../../lib/framework'
+import * as fs from 'fs'
+import { RUM_BUNDLE } from '../../lib/framework'
+import { APPLICATION_ID, CLIENT_TOKEN } from '../../lib/helpers/configuration'
+
 describe('Inject RUM with Puppeteer', () => {
   // S8s tests inject RUM with puppeteer evaluateOnNewDocument
   it('should not throw error in chrome', async () => {
@@ -6,3 +9,33 @@ describe('Inject RUM with Puppeteer', () => {
     expect(isInjected).toBe(true)
   })
 })
+
+async function injectRumWithPuppeteer() {
+  const ddRUM = fs.readFileSync(RUM_BUNDLE, 'utf8')
+  const puppeteerBrowser = await browser.getPuppeteer()
+  let injected = true
+
+  await browser.call(async () => {
+    const page = await puppeteerBrowser.newPage()
+    await page.evaluateOnNewDocument(
+      `
+        if (location.href !== 'about:blank') {
+          ${ddRUM}
+          window.DD_RUM._setDebug(true)
+          window.DD_RUM.init({
+            applicationId: ${APPLICATION_ID},
+            clientToken: ${CLIENT_TOKEN},
+          })
+          window.DD_RUM.startView()
+        }
+      `
+    )
+    page.on('console', (msg) => {
+      if (msg.type() === 'error') {
+        injected = false
+      }
+    })
+    await page.goto('https://webdriver.io')
+  })
+  return injected
+}
